+++ core/__init__.py
@@ -2,5 +2,5 @@
 
 Módulos:
     theme, ui, banner, parser, executor, response,
-    config, logger, history, update
+    config, logger, history, update, system
 """