+++ modules/mobile/__init__.py
@@ -1,6 +1,0 @@
-"""
-modules/mobile/
-
-Reservado para o aplicativo mobile e integração Desktop ↔ Mobile.
-Será implementado em versões futuras.
-"""